Add QuickCheck tests for Key and PrivKey
Adds QuickCheck tests for testing of basic operations for types which
are instances of the Key and PrivKey typeclasses. For now, Secp256k1
private keys are not tested as they require a particular message format
for use with verification and signing operations, which are more
difficult to generate automatically (at least with the current type
signature of the sign and verify functions).