Split Parse and Serialize, support only RSA Keys
Parsing of Ed25519 keys is ambiguous, because the 32 bytes used to
represent an Ed25519 Public Key could also be used to represent an
Ed25519 PrivateKey. Unfortunately the Protobuf serialization does not
distinguish between public and private keys, so we have no way of
knowing given the bytes of a private key whether it is a public or
private key.

There are also issues with Secp256k1 keys that need to be debugged.
Secp256k1 keys only support signing messages that have a specific format
in bytes.

For now, we disable Keys except RSA so that we can move to develop other
parts of LibP2P. The logic for serialization and parsing has also been
split into their own modules within libp2p-crypto.

Overall, I'm not currently satisfied with the model used to represent
keys, and there are a lot of improvements to be made. But this should
work at least for a first run so we can get PeerIds for LibP2P nodes.